Add Configure.cs

+++ LAB1/Object/Room.cs
@@ -44,9 +44,9 @@ namespace LAB1
         {
             List<Room> result = new List<Room>();
             Random rd = new Random();
-            for (uint i = 0; i < number_room; i++)
+            for (int i = 0; i < number_room; i++)
             {
-                int no = (int)i + 1;
+                int no = i + 1;
                 string id = Guid.NewGuid().ToString(P); // Generate random UUID V4 (32 bits) start with '{' and end with '}'
                 string classInfo = "N/A";
                 result.Add(new Room(id, classInfo, no));
